NotifyAndLog@NotificationManager
